Extracted -runTask method from -run method in AutoingestionJob.
Split the parts of -run that launch the NSTask and collect data from -run
into -runTask, which returns the task's output as a string.

Moved creation of the AutoingestionJob's description string into the init
method and added a description instance variable.  Overrode the -description
method to return the description string.

Removed the arugmentDateFormatter and descriptionDateFormatter properties
and replaced them with a local NSDateFormatter object that's only needed in
the init method.