Sleep and retry if UnknownHostException is thrown on first job.
If the first AutoingestionJob fails because the Auto-Ingest tool raises
an UnknownHostException, sleep for 60 seconds and try again.  This may
happen when the Mac is waking up and running delayed launchd jobs but the
network is slow to reconnect.

In AutoingestionJob, added enum AutoingestionResponseCode to give numeric
codes to various Auto-Ingest tool responses.  Added
-responseCodeFromResponse: method to AutoingestionJob to scan the output of
the Auto-Intest tool and determine the appropriate response code.  Added a
responseCode property to AutoingestionJob.

Localized creation and management of the NSTask object into the -runTask
method of AutoingestionJob.  Made task a local variable and removed the
task property.  Added an arguments property to hold the Auto-Ingest tool
argument list that's built in the init method of AutoingestionJob.

Modified the -downloadReports method of Tool to check the response code of
the first AutoingestionJob and to sleep and retry once if an
UnknownHostException occurs.