Filtered out Java exception stack trace lines from log messages.
For some kinds of common failures, the Autoingest tool prints out the stack
trace from the Java exception that is thrown.  These stack trace lines are
not useful, since Apple does not make the Java source code of the Autoingest
tool available.

Moved the code for processing the Autoingest tool's output out of the
AutoingestionJob class into the new AutoingestionResponse class.  The
AutoingestionResponse object examines the console output from the
Autoingest tool and determines if the job was successful.  It also creates
a summary of the output, removing any Java stack trace lines and indenting
lines after the first line with a tab.

Moved the AutoingestionResponseCode enum from AutoingestionJob to
AutoingestionResponse, renamed the responseCode property on
AutoingestionJob to response and made it an AutoingestionResponse object.
The -runTask method on AutoingestionJob now returns an AutoingestionResponse
object.  Also removed the -responseCodeFromResponse: method from
AutoingestionJob.

The AutoingestionResponse object has properties to hold the response code
enum, the downloaded filename for a successful job, the full response text,
including Java stack trace if present, and the response summary.  Convenience
properties "isSuccess" and "isNetworkUnavailable" calculate those conditions
based on the response code.

Modified AutoingestionJob -run and Tool -downloadReports to use the
"isSuccess" and "isNetworkUnavailable" methods respectively.

Added a UnitTests target with tests for the AutoingestionResponse class.

Added "local" to .gitignore.