Changed Group class to make a deep copy of its group struct.
Modified the -initWithGroup: designated initializer of the Group class so that
it makes a deep copy of the POSIX group struct that it is initialized with.
Added a -dealloc method to properly free the strings belonging to the deeply
copied group struct.

Added static helper methods duplicateNullTerminatedArrayOfStrings() and
freeNullTerminatedArrayOfStrings() to Group.m for copying the group members
list.

Added GroupTest to UnitTests covering the Group class.