Added installer plugin to automatically download the Auto-Ingest tool.
The Auto-Ingest tool is provided and maintained by Apple.  The
DownloadAutoingestionPlugin plugin adds an additional pane to the installer
that downloads the Auto-Ingest tool if it is not present.

The DownloadAutoingestionPluginPane class provides the UI for the installer
pane.  The AutoingestionTool class manages downloading and installing the
Auto-Ingest tool.  Currently, downloading is implemented, but the final
step, unzipping the downloaded .zip archive, is not implemented.