Implemented writing vendor credentials to Java .properties file.
The latest Auto-Ingest tool no longer accepts username and password on the
command line; in their place, it takes the path to a Java .properties file
containing "userID" and "password" keys.

Added a "path" property to VendorFile to expose the path of the vendor file.

Added "credentialsFilePath" property to Vendor class.  The credentialsFilePath
is created by dropping the .plist extension of the corresponding VendorFile
path and adding .properties.  Vendor's -prepare method now checks for the
credentials properties file and creates it if missing.

Changed initialization of AutoingestionJob to replace the username and
password arguments with the credentialsFilePath from the Vendor.