Made path to credentials .properties file relative rather than absolute.
The autoingestion jobs were failing with the message "Property file missing".
The AutoingestionJob class was creating the command line for the
Auto-Ingest tool using an absolute path to the .properties file with the
credentials; in manual testing of the Auto-Ingest tool I had used a relative
path.

Changed the init method of AutoingestionJob to build up a relative path
from the job's working directory (the directory where the particular report
being downloaded belongs) to the /Library/Autoindigestion/Vendors
directory.