Removed unnecessary ZipDocument instance variables.
The instance variables directoryEntriesStart and numberOfDirectoryEntries
were only holding temporary values needed while reading the zip archive.
Instead of storing them as part of the ZipDocument, added two similar
parameters to the -readEntries method, now named
-readEntriesWithDirectoryEntriesStart:andCount:.  The -readFromData:error:
method originates the two values and is the only caller of the "readEntries"
method.