Abort download after receiving a "try again" response.
Renamed enum "AutoingestionResponseCodeTryAgain" and related constant
to "...TryAgainLater".

Added "tryAgainLater" boolean property to AutoingestionResponse.
Added unit test coverage for "tryAgainLater" and "networkUnavailable"
properties.

Modified -downloadReports method in Tool to abort download when a
try again later response is received.

Background:

Changed Autoindigestion to stop downloading reports when the first
"try again in a few minutes" response is received from iTunes Connect.

Sometimes, iTunes Connect report generation is delayed and iTunes Connect
returns a "try again" response.  Previously, Autoindigestion checked for
the availability of each missing report in turn, even after the first
"try again" response, which can lead to skipping a report in rare
circumstances.  If, while checking for multiple reports of a single type
for a single vendor, the "try again" is received for the first report,
Autoindigestion would proceed to check for the next missing report in
chronological order.  It is possible that the newer reports might suddenly
become available and would be downloaded normally.  The next time that
Autoindigestion runs, it will look for the most recently downloaded report
for each vendor + report type, permanently skipping the earlier missed
report.

Here are sample logs illustrating the problem:

2013-08-08 07:47:16.512 Autoindigestion: 04-Aug-2013 Daily Sales Report for Able Pear Software: # this was skipped
	The report you requested is not available at this time.  Please try again in a few minutes.
2013-08-08 07:47:18.015 Autoindigestion: Downloaded 05-Aug-2013 Daily Sales Report for Able Pear Software: S_D_80082390_20130805.txt.gz
2013-08-08 07:47:18.734 Autoindigestion: Downloaded 06-Aug-2013 Daily Sales Report for Able Pear Software: S_D_80082390_20130806.txt.gz
2013-08-08 07:47:19.436 Autoindigestion: Downloaded 07-Aug-2013 Daily Sales Report for Able Pear Software: S_D_80082390_20130807.txt.gz

2013-08-10 07:30:18.150 Autoindigestion: Downloaded 08-Aug-2013 Daily Sales Report for Able Pear Software: S_D_80082390_20130808.txt.gz
2013-08-10 07:30:19.672 Autoindigestion: Downloaded 09-Aug-2013 Daily Sales Report for Able Pear Software: S_D_80082390_20130809.txt.gz