Don't try to download the yearly report for the current year.
Changed ReportCategory so that it does not generate an
AutoingestionJob for the current year for yearly reports.

Extracted -isValidReportDate: and -missingReportDates: methods
from the -autoingestionJobs method of ReportCategory, isolating the
building of report dates from reading the filesystem for existing
report files and generating autoingestion jobs.

Added some simple unit test coverate for -missingReportDates:.