Added -oldestReportDateBeforeDate: to ReportDateType.
Changed ReportDateType init method to take a block to use as the
implementation the -oldestReportDateBeforeDate: method.

Modified the implementation of some methods in the NSCalendar
(Autoindigestion) category to zero out the time parts when calculating
dates.  Renamed +posixDate to +POSIXDate.

In ReportCategory, implemented the -startingReportDate method in terms
of ReportDateType -oldestReportDateBeforeDate: and inlined it.