Improved download pane of installer.
Added controls for error message and skip/retry buttons.  Added a state
property to the DownloadAutoingestionPluginPane and reorganized how the
pane responds to various events.

Changed autoingestionTool:downloadFailedWithError: method in the
AutoingestionToolDelegate protocol to
autoingestionTool:downloadFailedWithMessage: so that it takes a string
instead of an NSError.

Modified AutoingestionTool to pass along failure messages for most error
conditions.